requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
appnope==0.1.2
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
decorator==5.1.0
idna==3.2
imageio==2.11.1
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
multidict==5.1.0
numpy==1.21.4
parso==0.8.2
peewee==3.14.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.21
ptyprocess==0.7.0
Pygments==2.10.0
python-dotenv==0.19.1
pytz==2021.1
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
yarl==1.6.3
